chore: 임시로 만든 Users와 UserRepository 삭제, domain/plant로 디렉토리 이름 변경
