+++ src/main/java/com/cookeep/cookeep/api/controller/PlantController.java
@@ -4,8 +4,8 @@ import com.cookeep.cookeep.api.dto.request.ProfilePlantRequest;
 import com.cookeep.cookeep.api.dto.response.MyPlantResponse;
 import com.cookeep.cookeep.api.dto.response.PlantResponse;
 import com.cookeep.cookeep.common.dto.DataResponse;
-import com.cookeep.cookeep.domain.Plant.application.PlantService;
-import com.cookeep.cookeep.domain.Plant.application.UserPlantService;
+import com.cookeep.cookeep.domain.plant.application.PlantService;
+import com.cookeep.cookeep.domain.plant.application.UserPlantService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
