+++ src/main/java/com/cookeep/cookeep/domain/plant/dao/PlantRepository.java
@@ -1,6 +1,6 @@
-package com.cookeep.cookeep.domain.Plant.dao;
+package com.cookeep.cookeep.domain.plant.dao;
 
-import com.cookeep.cookeep.domain.Plant.entity.Plant;
+import com.cookeep.cookeep.domain.plant.entity.Plant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
