+++ src/main/java/com/cookeep/cookeep/domain/user/dao/UserRepository.java
@@ -3,6 +3,8 @@ package com.cookeep.cookeep.domain.user.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cookeep.cookeep.domain.user.entity.User;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 }
